+++ insanic/conf/global_settings.py
@@ -148,9 +148,4 @@ INTERNAL_REQUEST_USER_HEADER = 'x-insanic-request-user'
 
 SERVICE_UNAVAILABLE_MESSAGE = "{} is currently unavailable."
 
-GRPC_SERVE = True
-GRPC_HOST = '0.0.0.0'
-GRPC_PORT_DELTA = 1000
-# needs to be something like "{service_name}.{server_class}"
-GRPC_SERVER = []