+++ insanic/log/__init__.py
@@ -26,12 +26,6 @@ def get_logging_config():
                 "handlers": ["access_console"],
                 "propagate": True,
                 "qualname": "sanic.access"
-            },
-            "sanic.grpc": {
-                "level": 'INFO',
-                "handlers": ["access_grpc_console"],
-                "propagate": True,
-                "qualname": "sanic.grpc"
             }
         },
         handlers={
@@ -49,12 +43,7 @@ def get_logging_config():
                 "class": "logging.StreamHandler",
                 "formatter": "json",
                 "stream": sys.stdout
-            },
-            "access_grpc_console": {
-                "class": "logging.StreamHandler",
-                "formatter": "json_grpc",
-                "stream": sys.stdout
-            },
+            }
         },
         formatters={
             "generic": {
@@ -81,7 +70,6 @@ def get_logging_config():
                            "sampled": "%(ot_sampled)s", "request_duration": "%(request_duration)s",
                            "parent_id": "%(ot_parent_id)s", "correlation_id": "%(correlation_id)s",
                            "exc_text": "%(exc_text)s", "request_service": "%(request_service)s",
-                           "is_grpc": "%(is_grpc)d",
                            # added in 0.6.8
                            "error_code_name": "%(error_code_name)s",
                            "error_code_value": "%(error_code_value)s",
@@ -90,40 +78,6 @@ def get_logging_config():
                            "uri_template": "%(uri_template)s",
                            # added in 0.8.0
                            "squad": "%(squad)s"
-                           },
-                'datefmt': '%Y-%m-%dT%H:%M:%S.%%(msecs)d%z'
-            },
-            "access_grpc": {
-                "format": "%(asctime)s - (%(name)s)[%(levelname)s][%(host)s]: " +
-                          "GRPC %(cardinality)s %(scheme)s://%(host)s%(path)s %(message)s %(status)s %(grpc_status)s",
-                "datefmt": "[%Y-%m-%d %H:%M:%S %z]",
-                "class": "logging.Formatter"
-            },
-            "json_grpc": {
-                "()": "insanic.log.formatters.JSONFormatter",
-                "format": {"level": "%(levelname)s", "hostname": "%(hostname)s", "where": "%(module)s.%(funcName)s",
-                           "ts": "%(asctime)s", "message": "%(message)s",
-                           "status": "%(status)s", "name": "%(name)s",
-                           "service": "%(service)s",
-                           "environment": "%(environment)s", "insanic_version": "%(insanic_version)s",
-                           "service_version": "%(service_version)s",
-                           "correlation_id": "%(correlation_id)s",
-                           "exc_text": "%(exc_text)s", "request_service": "%(request_service)s",
-                           # added in 0.6.8
-                           "method": "%(method)s",
-                           "path": "%(path)s",
-                           "grpc_status": "%(grpc_status)s",
-                           "stream_id": "%(stream_id)s",
-                           "cardinality": "%(cardinality)s",
-                           # added in 0.8.0
-                           "squad": "%(squad)s"
-
-                           # "size": "%(byte)d",
-                           # "trace_id": "%(ot_trace_id)s",
-                           # "span_id": "%(ot_span_id)s",
-                           # "sampled": "%(ot_sampled)s",
-                           # "request_duration": "%(request_duration)s",
-                           # "parent_id": "%(ot_parent_id)s",
                            },
                 'datefmt': '%Y-%m-%dT%H:%M:%S.%%(msecs)d%z'
             }
@@ -135,18 +89,13 @@ def get_logging_config():
         LOGGING_CONFIG_DEFAULTS['loggers']['root']['level'] = logging.DEBUG
         LOGGING_CONFIG_DEFAULTS['loggers']['sanic.error']['level'] = logging.DEBUG
         LOGGING_CONFIG_DEFAULTS['loggers']['sanic.access']['level'] = logging.DEBUG
-        LOGGING_CONFIG_DEFAULTS['loggers']['sanic.grpc']['level'] = logging.DEBUG
         LOGGING_CONFIG_DEFAULTS['handlers']['console']['formatter'] = 'generic'
         LOGGING_CONFIG_DEFAULTS['handlers']['error_console']['formatter'] = 'generic'
         LOGGING_CONFIG_DEFAULTS['handlers']['access_console']['formatter'] = 'access'
-        LOGGING_CONFIG_DEFAULTS['handlers']['access_grpc_console']['formatter'] = 'access_grpc'
 
     return LOGGING_CONFIG_DEFAULTS
 
 
 logger = logging.getLogger('root')
-grpc_logger = logging.getLogger('root.grpc')
 error_logger = logging.getLogger('sanic.error')
-grpc_error_logger = logging.getLogger('sanic.error.grpc')
 access_logger = logging.getLogger('sanic.access')
-grpc_access_logger = logging.getLogger('sanic.grpc')